+++ CHANGELOG.md
@@ -9,10 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+
 - `shroom proxy` command: Proxy & translate WebSocket messages to something the emulator will understand
   This enables you to communicate with the emulator through websockets, without making any adjustments to
   the emulator itself.
 - Option to hide tile cursor in a room with `room.hideTileCursor`
+- BaseFurniture now has a onLoad callback, which gets called after the furniture is fully loaded
 
 ### Fixed